refactor(guest-agent): extract cli termination fsm
